removed nuxt ui components
