+++ app/pages/contact.vue
@@ -1,19 +1,13 @@
 <script setup lang="ts">
 const { data: contact } = await useAsyncData(() =>
-	queryCollection("content").path("/contact").first(),
+  queryCollection("content").path("/contact").first(),
 );
 
 useSeoMeta({
-	title: contact.value?.title,
-	description: contact.value?.description,
+  title: contact.value?.title,
+  description: contact.value?.description,
 });
 </script>
 <template>
-  <UPageBody>
-    <ContentRenderer
-      v-if="contact"
-      :value="contact"
-      class="flow content-grid"
-    />
-  </UPageBody>
+  <ContentRenderer v-if="contact" :value="contact" class="flow content-grid" />
 </template>
