+++ app/pages/teachers/[...slug].vue
@@ -12,6 +12,8 @@ useSeoMeta({
 </script>
 
 <template>
-  <NuxtImg :src="data?.meta.image" width="300" height="400" format="avif" />
-  <ContentRenderer v-if="data" :value="data" />
+	<div>
+		<NuxtImg :src="data?.meta.image" width="300" height="400" format="avif" />
+		<ContentRenderer v-if="data" :value="data" />
+	</div>
 </template>
